+++ CHANGES.md
@@ -7,6 +7,8 @@ changes.
 
 ## Unreleased
 
+- Always start in snap mode for a new route, even if the user last used freehand mode.
+
 ## 0.2.2
 
 - Change the `renderGeojson` output to distinguish snapped and freehand waypoints